backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
click==8.3.3
fastapi==0.136.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
joblib==1.5.3
numpy==2.4.4
packaging==26.2
pandas==3.0.2
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
pygments==2.20.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-multipart==0.0.27
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
SQLAlchemy==2.0.43
starlette==1.0.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
asyncpg==0.31.0
psycopg2-binary==2.9.11
